+++ package.json
@@ -25,8 +25,8 @@
     "tailwindcss": "^3.4.17"
   },
   "dependencies": {
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19",
+    "react-dom": "19",
     "react-router-dom": "^6.28.1"
   },
   "browserslist": [
